rename enums folder to Enum
